feat: naver analytics 설정, 분기 처리
