+++ components/analytics/NaverAnalytics.tsx
@@ -1,0 +1,26 @@
+import Script from "next/script";
+import * as naver from '@/lib/naver';
+
+export const NaverAnalytics = () => {
+  return (
+    <>
+      <Script
+        strategy="afterInteractive"
+        src="//wcs.naver.net/wcslog.js"
+      ></Script>
+      <Script
+        id="naver-init"
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `
+              if(!wcs_add) var wcs_add = {};
+              wcs_add["wa"] = "${naver.NAVER_TRACKING_ID}";
+              if(window.wcs) {
+                wcs_do();
+              }
+            `,
+        }}
+      />
+    </>
+  );
+};
