+++ pages/_app.tsx
@@ -1,44 +1,15 @@
+import { GoogleAnalytics } from '@/components/analytics/GoogleAnalytics';
+import { NaverAnalytics } from '@/components/analytics/NaverAnalytics';
 import { ThemeProvider } from 'next-themes';
 import type { AppProps } from 'next/app';
-import { useRouter } from 'next/router';
-import Script from 'next/script';
-import React, { useEffect } from 'react';
+import React from 'react';
 import '../styles/globals.css';
-import * as gtag from '../lib/gtag';
 
 const MyApp = ({ Component, pageProps }: AppProps): JSX.Element => {
-  const router = useRouter();
-  useEffect(() => {
-    const handleRouteChange = (url: URL) => {
-      gtag.pageview(url);
-    };
-    router.events.on('routeChangeComplete', handleRouteChange);
-    router.events.on('hashChangeComplete', handleRouteChange);
-    return () => {
-      router.events.off('routeChangeComplete', handleRouteChange);
-      router.events.off('hashChangeComplete', handleRouteChange);
-    };
-  }, [router.events]);
   return (
     <ThemeProvider attribute="class" enableSystem={false} defaultTheme="light">
-      <Script
-        strategy="afterInteractive"
-        src={`https://www.googletagmanager.com/gtag/js?id=${gtag.GA_TRACKING_ID}`}
-      />
-      <Script
-        id="gtag-init"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-              window.dataLayer = window.dataLayer || [];
-              function gtag(){dataLayer.push(arguments);}
-              gtag('js', new Date());
-              gtag('config', '${gtag.GA_TRACKING_ID}', {
-                page_path: window.location.pathname,
-              });
-            `,
-        }}
-      />
+      <GoogleAnalytics/>
+      <NaverAnalytics/>
       <Component {...pageProps} />
     </ThemeProvider>
   );
